+++ shared/framework-kernel/src/declarations/index.ts
@@ -1,0 +1,13 @@
+import type { Router, RouteLocationRaw } from 'vue-router';
+
+export interface RouterOptions {
+  instance: Router;
+  path: {
+    root: RouteLocationRaw;
+    home: RouteLocationRaw;
+    signIn: RouteLocationRaw;
+  };
+}
+export interface KernelOptions {
+  router: RouterOptions;
+}