+++ shared/framework-kernel/src/lib/base/index.ts
@@ -1,0 +1,1 @@
+export { lodash, Toolkit } from '@herodotus-cloud/core';