+++ shared/apis/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @herodotus-cloud/apis
 
+## 1.6.1
+
+### Patch Changes
+
+- 修复Router单例
+- Updated dependencies
+  - @herodotus-cloud/core@1.6.1
+
+## 1.6.0
+
+### Minor Changes
+
+- 新增 framework kernel 模块，提取框架公共类以支持不同框架
+
+### Patch Changes
+
+- Updated dependencies
+  - @herodotus-cloud/core@1.6.0
+
 ## 1.5.1
 
 ### Patch Changes