+++ shared/form-designer/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @herodotus-cloud/form-designer
 
+## 1.6.1
+
+### Patch Changes
+
+- 修复Router单例
+- Updated dependencies
+  - @herodotus-cloud/form-engine@1.6.1
+  - @herodotus-cloud/components@1.6.1
+  - @herodotus-cloud/form-apis@1.6.1
+  - @herodotus-cloud/core@1.6.1
+
+## 1.6.0
+
+### Minor Changes
+
+- 新增 framework kernel 模块，提取框架公共类以支持不同框架
+
+### Patch Changes
+
+- Updated dependencies
+  - @herodotus-cloud/form-engine@1.6.0
+  - @herodotus-cloud/components@1.6.0
+  - @herodotus-cloud/form-apis@1.6.0
+  - @herodotus-cloud/core@1.6.0
+
 ## 1.5.1
 
 ### Patch Changes